+++ R/internals.R
@@ -34,6 +34,17 @@ fingerprint <- function(x){
   }
 } # end function
 
+#' Run fingerprint on a dataframe
+#'
+#' @param x data frame
+#'
+#' @return data frame with column types and NA values
+#' @noRd
+#' @note internal
+fingerprint_df <- function(x){
+  purrr::map_df(x, fingerprint)
+}
+
 
 #' (Internal) Gather rows into a format appropriate for grid visualisation
 #'
@@ -46,7 +57,7 @@ fingerprint <- function(x){
 #'
 vis_gather_ <- function(x){
   x %>%
-  dplyr::mutate(rows = dplyr::row_number()) %>%
+    dplyr::mutate(rows = dplyr::row_number()) %>%
     tidyr::pivot_longer(
       cols = -rows,
       names_to = "variable",
@@ -107,7 +118,8 @@ vis_create_ <- function(x){
                 y = "Observations") +
     # flip the axes
     ggplot2::scale_y_reverse() +
-    ggplot2::theme(axis.text.x = ggplot2::element_text(hjust = 0.5))
+    ggplot2::theme(axis.text.x = ggplot2::element_text(hjust = 0.5)) +
+    ggplot2::guides(colour = "none")
 
 }
 
@@ -133,66 +145,71 @@ vis_create_ <- function(x){
 add_vis_dat_pal <- function(vis_plot, palette){
 
   # palette options: http://docs.ggplot2.org/current/discrete_scale.html
-# qualitative, 6 colours --------------------------------------------------
-vis_pal_qual <- c("#e41a1c", # red
-                      "#ffff33", # yellow
-                      "#ff7f00", # Orange
-                      "#377eb8", # blue
-                      "#4daf4a", # Green
-                      "#984ea3") # Purple
+  # qualitative, 6 colours --------------------------------------------------
+  vis_pal_qual <- c("#e41a1c", # red
+                             "#ffff33", # yellow
+                             "#ff7f00", # Orange
+                             "#377eb8", # blue
+                             "#4daf4a", # Green
+                             "#984ea3") # Purple
 
-# diverging, 6 colours, colour-blind safe -------------------------------
-vis_pal_cb_safe <- c('#d73027', # red
-                         '#fc8d59', # orange
-                         '#fee090', # yellow
-                         '#e0f3f8', # light blue
-                         '#91bfdb', # mid blue
-                         '#4575b4') # dark blue
+  # diverging, 6 colours, colour-blind safe -------------------------------
+  vis_pal_cb_safe <- c('#d73027', # red
+                                '#fc8d59', # orange
+                                '#fee090', # yellow
+                                '#e0f3f8', # light blue
+                                '#91bfdb', # mid blue
+                                '#4575b4') # dark blue
 
-if (palette == "default"){
+  if (palette == "default"){
 
-  vis_plot
+    vis_plot
 
-} else if (palette == "qual") {
+  } else if (palette == "qual") {
 
-  vis_plot +
-    ggplot2::scale_fill_manual(limits = c("character",
-                                          "date",
-                                          "factor",
-                                          "integer",
-                                          "logical",
-                                          "numeric"),
-                               breaks = c("character", # red
-                                          "date", # orange
-                                          "factor", # yellow
-                                          "integer", # light blue
-                                          "logical", # mid blue
-                                          "numeric"), # dark blue
-                               values = vis_pal_qual,
-                               na.value = "grey")
+    vis_plot +
+      ggplot2::scale_fill_manual(limits = c("character",
+                                            "date",
+                                            "factor",
+                                            "integer",
+                                            "logical",
+                                            "numeric"),
+                                 breaks = c("character", # red
+                                            "date", # orange
+                                            "factor", # yellow
+                                            "integer", # light blue
+                                            "logical", # mid blue
+                                            "numeric"), # dark blue
+                                 values = vis_pal_qual,
+                                 na.value = "grey")
 
 
-} else if (palette == "cb_safe") {
+  } else if (palette == "cb_safe") {
 
-  vis_plot +
-    ggplot2::scale_fill_manual(limits = c("character",
-                                          "date",
-                                          "factor",
-                                          "integer",
-                                          "logical",
-                                          "numeric"),
-                               breaks = c("character", # red
-                                          "date", # orange
-                                          "factor", # yellow
-                                          "integer", # light blue
-                                          "logical", # mid blue
-                                          "numeric"), # dark blue
-                               values = vis_pal_cb_safe,
-                               na.value = "grey")
+    vis_plot +
+      ggplot2::scale_fill_manual(limits = c("character",
+                                            "date",
+                                            "factor",
+                                            "integer",
+                                            "logical",
+                                            "numeric"),
+                                 breaks = c("character", # red
+                                            "date", # orange
+                                            "factor", # yellow
+                                            "integer", # light blue
+                                            "logical", # mid blue
+                                            "numeric"), # dark blue
+                                 values = vis_pal_cb_safe,
+                                 na.value = "grey")
 
-} else  {
-  stop("palette arguments need to be either 'qual' 'cb_safe' or 'default'")
-} # close else brace
+  } else  {
+    cli::cli_abort(
+      c(
+        "Palette arguments need to be one of: 'qual', 'cb_safe', or 'default'",
+        "You palette argument was: {.arg {palette}}"
+      )
+    )
+  } # close else brace
 
 } # close the function
 
@@ -210,17 +227,16 @@ label_col_missing_pct <- function(x,
 
   # present everything in the right order
 
-  labelled_pcts <-
-  purrr::map_df(x, ~round(mean(is.na(.))*100,2))[col_order_index] %>%
+  labelled_pcts <- colMeans(is.na(x))[col_order_index] %>%
     purrr::map_chr(function(x){
       dplyr::case_when(
         x == 0 ~  "0%",
-        x >= 0.1 ~ as.character(glue::glue("{x}%")),
-        x < 0.1 ~ "<0.1%"
+        x >= 0.001 ~ scales::percent(x, accuracy = 1),
+        x < 0.001 ~ "<0.1%"
       )
     })
 
-    glue::glue("{col_order_index} ({labelled_pcts})")
+  glue::glue("{col_order_index} ({labelled_pcts})")
 
 }
 
@@ -264,7 +280,7 @@ miss_guide_label <- function(x) {
 
     p_pres_lab <- glue::glue("Present \n({p_pres}%)")
 
-    }
+  }
 
   label_frame <- tibble::tibble(p_miss_lab,
                                 p_pres_lab)
@@ -319,56 +335,40 @@ all_binary <- function(x, ...){
 #' @noRd
 test_if_dataframe <- function(x){
 
-  msg <- cli::format_error(
-    c(
-      "{.code vis_dat()} requires a {.cls data.frame}",
-      "the object I see has class(es): ",
-      "{.cls {glue::glue_collapse(class(x), sep = ', ', last = ', and ')}}"
-    )
-  )
-
   if (!inherits(x, "data.frame")) {
-    stop(
-      msg,
-      call. = FALSE)
-
+    cli::cli_abort(
+      c(
+        "{.code vis_dat()} requires a {.cls data.frame}",
+        "the object I see has class(es): ",
+        "{.cls {glue::glue_collapse(class(x), sep = ', ', last = ', and ')}}"
+      )
+    )
   }
 }
 
 test_if_all_numeric <- function(data){
-
-  msg <- cli::format_error(
-    c(
-      "Data input can only contain numeric values",
-      "Please subset the data to the numeric values you would like.",
-      "{.code dplyr::select(<data>, where(is.numeric))}",
-      "Can be helpful here!"
-    )
-  )
-
   if (!all_numeric(data)) {
-    stop(
-      msg,
-      call. = FALSE
+    cli::cli_abort(
+      c(
+        "Data input can only contain numeric values",
+        "Please subset the data to the numeric values you would like.",
+        "{.code dplyr::select(<data>, where(is.numeric))}",
+        "Can be helpful here!"
+      )
     )
-    }
   }
+}
 
 test_if_all_binary <- function(data){
 
-  msg <- cli::format_error(
-    c(
-      "data input can only contain binary values",
-      "This means values are either 0 or 1, or NA.",
-      "Please subset the data to be binary values, or see {.code ?vis_value.}"
-    )
-  )
-
   if (!all_binary(data)) {
-    stop(
-      msg,
-      call. = FALSE
+    cli::cli_abort(
+      c(
+        "data input can only contain binary values",
+        "This means values are either 0 or 1, or NA.",
+        "Please subset the data to be binary values, or see {.code ?vis_value.}"
       )
+    )
   }
 }
 
@@ -386,4 +386,57 @@ scale_01 <- function(x) {
     return((x/x))
   }
   (x - min(x, na.rm = TRUE)) / diff(range(x, na.rm = TRUE))
+}
+
+group_by_fun <- function(data,.fun, ...){
+  tidyr::nest(data) %>%
+    dplyr::mutate(data = purrr::map(data, .fun, ...)) %>%
+    tidyr::unnest(cols = c(data))
+}
+
+data_vis_class_not_implemented <- function(fun){
+  cli::cli_abort(
+    c(
+      "We have not (yet) implemented the method for {.code fun} for \\
+      object with class(es): ",
+      "{.cls {glue::glue_collapse(class(x), sep = ', ', last = ', and ')}}"
+    )
+  )
+}
+
+update_col_order_index <- function(col_order_index, facet, env = environment()){
+  group_string <- deparse(substitute(facet, env = env))
+
+  facet_position <- which(col_order_index == group_string)
+  col_order_index <- col_order_index[-facet_position]
+}
+
+test_if_large_data <- function(x, large_data_size, warn_large_data){
+  if (ncol(x) * nrow(x) > large_data_size && warn_large_data){
+    cli::cli_abort(
+      c(
+        "Data exceeds recommended size for visualisation",
+        "Consider downsampling your data with {.fn dplyr::slice_sample}",
+        "Or set argument, {.arg warn_large_data} = {.arg FALSE}"
+      )
+    )
+  }
+}
+
+fast_n_miss_col <- function(x) colSums(is.na(x))
+
+n_miss_col <- function(data, sort = FALSE){
+  # if no list columns
+  any_list <- any(purrr::map_lgl(data, is.list))
+  if (!any_list){
+    n_missing_cols <- fast_n_miss_col(data)
+  } else if (any_list){
+    n_missing_cols <- fast_n_miss_col(fingerprint_df(data))
+  }
+
+ if (sort){
+   n_missing_cols <- sort(n_missing_cols, decreasing = TRUE)
+ }
+
+  n_missing_cols
 }