+++ R/vis-miss.R
@@ -5,6 +5,10 @@
 #'   a missing cell and grey indicates a present cell. As it returns a ggplot
 #'   object, it is very easy to customize and change labels.
 #'
+#' The missingness summaries in the columns are rounded to the nearest integer.
+#'   For more detailed summaries, please see the summaries in the `naniar` R
+#'   package, specifically, `naniar::miss_var_summary()`.
+#'
 #' @param x a data.frame
 #'
 #' @param cluster logical. TRUE specifies that you want to use hierarchical
@@ -17,8 +21,11 @@
 #' @param show_perc logical. TRUE now adds in the \% of missing/complete data
 #'   in the whole dataset into the legend. Default value is TRUE.
 #'
-#'@param show_perc_col logical. TRUE adds in the \% missing data in a given
+#' @param show_perc_col logical. TRUE adds in the \% missing data in a given
 #'  column into the x axis. Can be disabled with FALSE. Default value is TRUE.
+#'  No missingness percentage column information will be presented when `facet`
+#'  argument is used. Please see the `naniar` package to provide missingness
+#'  summaries over groups.
 #'
 #' @param warn_large_data logical - warn if there is large data? Default is TRUE
 #'   see note for more details
@@ -26,6 +33,11 @@
 #' @param large_data_size integer default is 900000 (given by
 #'   `nrow(data.frame) * ncol(data.frame)``). This can be changed. See
 #'   note for more details.
+#'
+#' @param facet (optional) bare variable name, if you want to create a faceted
+#'   plot, with one plot per level of the variable. No missingness percentage
+#'   column information will be presented when `facet` argument is used. Please
+#'   see the `naniar` package to provide missingness summaries over groups.
 #'
 #' @return `ggplot2` object displaying the position of missing values in the
 #'   dataframe, and the percentage of values missing and present.
@@ -46,7 +58,9 @@
 #'
 #' vis_miss(airquality, sort_miss = TRUE)
 #'
-#'\dontrun{
+#' vis_miss(airquality, facet = Month)
+#'
+#' \dontrun{
 #' # if you have a large dataset, you might want to try downsampling:
 #' library(nycflights13)
 #' library(dplyr)
@@ -57,83 +71,51 @@
 #' flights %>%
 #'   slice(1:1000) %>%
 #'   vis_miss()
-#'}
+#' }
 #'
 #' @export
-vis_miss <- function(x,
-                     cluster = FALSE,
-                     sort_miss = FALSE,
-                     show_perc = TRUE,
-                     show_perc_col = TRUE,
-                     large_data_size = 900000,
-                     warn_large_data = TRUE){
+vis_miss <- function(
+  x,
+  cluster = FALSE,
+  sort_miss = FALSE,
+  show_perc = TRUE,
+  show_perc_col = TRUE,
+  large_data_size = 900000,
+  warn_large_data = TRUE,
+  facet
+    ) {
 
-  # throw error if x not data.frame
   test_if_dataframe(x)
-
-  # add warning for large data
-  if (ncol(x) * nrow(x) > large_data_size && warn_large_data) {
-    stop("Data exceeds recommended size for visualisation, please consider
-         downsampling your data, or set argument 'warn_large_data' to FALSE.")
-  }
-
-  # make a TRUE/FALSE matrix of the data.
-  # This tells us whether it is missing (true) or not (false)
-  x.fingerprinted <- x %>%
-    purrr::map_df(fingerprint)
-
-  x.na <- x %>%
-    purrr::map_df(~fingerprint(.x) %>% is.na)
-
-  # switch for creating the missing clustering
-  if (cluster){
-
-    # this retrieves a row order of the clustered missingness
-    row_order_index <-
-      stats::dist(x.na*1) %>%
-      stats::hclust(method = "mcquitty") %>%
-      stats::as.dendrogram() %>%
-      stats::order.dendrogram()
-
-  } else {
-    row_order_index <- seq_len(nrow(x))
-  } # end else
+  test_if_large_data(x, large_data_size, warn_large_data)
 
   if (sort_miss) {
-
-    # arrange by the columns with the highest missingness
-    # code inspired from https://r-forge.r-project.org/scm/viewvc.php/ ...
-    # pkg/R/missing.pattern.plot.R?view=markup&root=mi-dev
-    # get the order of columns with highest missingness
-    na_sort <- order(colSums(x.na), decreasing = TRUE)
-
-    # get the names of those columns
-    col_order_index <- names(x)[na_sort]
-
+    col_order_index <- names(n_miss_col(x, sort = TRUE))
   } else if (!sort_miss) {
-
     col_order_index <- names(x)
-
   }
 
-  # Arranged data by dendrogram order index
+  if (!missing(facet)) {
+    vis_miss_data <- x %>%
+      dplyr::group_by({{ facet }}) %>%
+      data_vis_miss(cluster)
 
-  # gather the variables together for plotting
-  # here we now have a column of the row number (row),
-  # then the variable(variables),
-  # then the contents of that variable (value)
-  dat_pre_vis <- as.data.frame(x.na[row_order_index , ])
+    col_order_index <- update_col_order_index(
+      col_order_index,
+      facet,
+      environment()
+      )
 
-  d <- dat_pre_vis %>%
-    vis_gather_() %>%
-    # add info for plotly mousover
-    dplyr::mutate(value = vis_extract_value_(dat_pre_vis))
+  } else {
+    vis_miss_data <- data_vis_miss(x, cluster)
+  }
 
   # calculate the overall % missingness to display in legend -------------------
+  # make a TRUE/FALSE matrix of the data.
+  # This tells us whether it is missing (true) or not (false)
+  x_fingerprinted <- fingerprint_df(x)
 
   if (show_perc) {
-
-    temp <- miss_guide_label(x.fingerprinted)
+    temp <- miss_guide_label(x_fingerprinted)
 
     p_miss_lab <- temp$p_miss_lab
 
@@ -141,19 +123,24 @@ vis_miss <- function(x,
 
     # else if show_perc FALSE
   } else {
-
     p_miss_lab <- "Missing"
 
     p_pres_lab <- "Present"
-
   }
+
   # then we plot it
-  vis_miss_plot <- vis_create_(d) +
-    ggplot2::scale_fill_manual(name = "",
-                               values = c("grey80",
-                                          "grey20"),
-                               labels = c(p_pres_lab,
-                                          p_miss_lab)) +
+  vis_miss_plot <- vis_create_(vis_miss_data) +
+    ggplot2::scale_fill_manual(
+      name = "",
+      values = c(
+        "grey80",
+        "grey20"
+      ),
+      labels = c(
+        p_pres_lab,
+        p_miss_lab
+      )
+    ) +
     ggplot2::guides(fill = ggplot2::guide_legend(reverse = TRUE)) +
     ggplot2::theme(legend.position = "bottom") +
     # fix up the location of the text
@@ -165,50 +152,56 @@ vis_miss <- function(x,
   # this is perhaps a bit of a hacky way around, but I can't see another
   # way around it. Related issue: https://github.com/ropensci/visdat/issues/72
   if (ncol(x) == 1) {
-
     if (show_perc_col) {
       return(
-        # print(
-      vis_miss_plot +
-        ggplot2::scale_x_discrete(position = "top",
-                                  labels = label_col_missing_pct(
-                                    x.fingerprinted,
-                                    col_order_index)
-        )
-      # )
+        vis_miss_plot <- vis_miss_plot +
+          ggplot2::scale_x_discrete(
+            position = "top",
+            labels = label_col_missing_pct(
+              x_fingerprinted,
+              col_order_index
+            )
+          )
       )
     } else if (!show_perc_col) {
       return(
-        # print(
-      vis_miss_plot +
-        ggplot2::scale_x_discrete(position = "top",
-                                  labels = col_order_index)
-      # )
+        vis_miss_plot <- vis_miss_plot +
+          ggplot2::scale_x_discrete(
+            position = "top",
+            labels = col_order_index
+          )
       )
     }
-
   }
 
-  if (show_perc_col) {
+  if (!missing(facet)) {
+    vis_miss_plot <- vis_miss_plot +
+      ggplot2::facet_wrap(facets = dplyr::vars({{ facet }}))
+  }
 
+  if (show_perc_col && missing(facet)) {
     # flip the axes, add the info about limits
-    vis_miss_plot +
-      ggplot2::scale_x_discrete(position = "top",
-                                limits = col_order_index,
-                                labels = label_col_missing_pct(
-                                  x.fingerprinted,
-                                  col_order_index)
+    vis_miss_plot <- vis_miss_plot +
+      ggplot2::scale_x_discrete(
+        position = "top",
+        limits = col_order_index,
+        labels = label_col_missing_pct(
+          x_fingerprinted,
+          col_order_index
+        )
       )
-
   } else {
-    vis_miss_plot +
-      ggplot2::scale_x_discrete(position = "top",
-                                limits = col_order_index)
+    vis_miss_plot <- vis_miss_plot +
+      ggplot2::scale_x_discrete(
+        position = "top",
+        limits = col_order_index
+      )
   }
+
+  return(vis_miss_plot)
 
   # guides(fill = guide_legend(title = "Type"))
   # Thanks to
   # http://www.markhneedham.com/blog/2015/02/27/rggplot-controlling-x-axis-order/
   # For the tip on using scale_x_discrete
-
-  } # end of function
+} # end of function